+++ src/directives/page/page.component.ts
@@ -20,6 +20,7 @@ export class WaPageComponent implements OnChanges {
   @Input() navOpen?: boolean;
   @Input() disableSticky?: string;
   @Input() navigationPlacement?: 'start' | 'end';
+  @Input() disableNavigationToggle?: boolean | string;
 
   @Input() menuWidth?: string;
   @Input() mainWidth?: string;
@@ -50,6 +51,11 @@ export class WaPageComponent implements OnChanges {
     }
     if (this.navigationPlacement !== undefined) {
       el.setAttribute('navigation-placement', this.navigationPlacement);
+    }
+    if (this.disableNavigationToggle === true || this.disableNavigationToggle === 'true' || this.disableNavigationToggle === '') {
+      el.setAttribute('disable-navigation-toggle', '');
+    } else if (this.disableNavigationToggle === false || this.disableNavigationToggle === 'false') {
+      el.removeAttribute('disable-navigation-toggle');
     }
 
     this.setCssVar('--menu-width', this.menuWidth || '');